Split track data into multiple object files
This original goal of this commit was to reduce the time required to
compile all of the tracks. While it now takes *longer* on my Core i7 Duo,
my CPUs are now actually maxed at 100%, and disk I/O is fully saturated!

An added benefit, is that this should alleviate memory related crashes in
Brandon's compiler, as the compiler only has to deal w/ a handful of ~4MiB
and ~70MiB source files instead of a single ~200MiB source file.